+++ src/hkimjp/wil2/util.clj
@@ -1,4 +1,8 @@
-(ns hkimjp.wil2.util)
+(ns hkimjp.wil2.util
+  (:require [java-time.api :as jt]))
 
 (defn user [request]
   (get-in request [:session :identity]))
+
+(defn today []
+  (str (jt/local-date)))