+++ src/hkimjp/wil2/util.clj
@@ -1,4 +1,16 @@
-(ns hkimjp.wil2.util)
+(ns hkimjp.wil2.util
+  (:require
+   [clojure.string :as str]
+   [java-time.api :as jt]))
 
 (defn user [request]
   (get-in request [:session :identity]))
+
+(defn today []
+  (str (jt/local-date)))
+
+(defn abbrev
+  "shorten string for concise log."
+  ([s] (abbrev s 80))
+  ([s n] (let [pat (re-pattern (str "(^.{" n "}).*"))]
+           (str/replace-first s pat "$1..."))))