+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - no effect reload/reload. how to use?
 - test
 - kill just nrepl by C-c, tailwindcss survives.
-
+- default-handler works?
 
 
 # 0.2.3-SNAPSHOT (2025-09-04)