+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 - no effect reload/reload. how to use?
 - test
 - kill just nrepl by C-c, tailwindcss survives.
+- default-handler works?
+
+
+# 0.2.3-SNAPSHOT (2025-09-04)
+
+- help - mx-2
+- upload - m-4
+- todays/upload - upload markdown files.
+    - in form: enctype="multipart/form-data"
+    - htmx: :hx-encoding "multipart/form-data"
 
 
 # 0.2.2 (2025-09-04)